+++ R/generate_pseudo_pop.R
@@ -144,6 +144,12 @@ generate_pseudo_pop <- function(Y,
   covariate_cols <- as.list(colnames(c))
 
 
+  # Depreciation messages
+  if (!optimized_compile){
+    warning("optimized_compile = FALSE will be depreciated.",
+            call. = FALSE)
+  }
+
   # get trim quantiles and trim data
   q1 <- stats::quantile(w,trim_quantiles[1])
   q2 <- stats::quantile(w,trim_quantiles[2])
@@ -242,6 +248,16 @@ generate_pseudo_pop <- function(Y,
                                              nthread = nthread,
                                              optimized_compile = optimized_compile,
                                              ...)
+    # check Kolmogorov-Smirnov statistics
+    ks_stats <- check_kolmogorov_smirnov(w = pseudo_pop[, c("w")],
+                                         c = pseudo_pop[,
+                                                        unlist(covariate_cols),
+                                                        with = FALSE],
+                                         counter_weight = pseudo_pop[,
+                                                           c("counter_weight")],
+                                         ci_appr = ci_appr,
+                                         nthread = nthread,
+                                         optimized_compile = optimized_compile)
 
     covar_bl_t <- paste0(covar_bl_trs_type,"_absolute_corr")
     if (is.null(best_ach_covar_balance)){
@@ -249,6 +265,7 @@ generate_pseudo_pop <- function(Y,
       best_pseudo_pop <- pseudo_pop
       best_adjusted_corr_obj <- adjusted_corr_obj
       best_gps_used_params <- gps_used_params
+      best_ks_stats <- ks_stats
     }
 
     if (getElement(adjusted_corr_obj$corr_results,covar_bl_t) < best_ach_covar_balance){
@@ -256,6 +273,7 @@ generate_pseudo_pop <- function(Y,
       best_pseudo_pop <- pseudo_pop
       best_adjusted_corr_obj <- adjusted_corr_obj
       best_gps_used_params <- gps_used_params
+      best_ks_stats <- ks_stats
     }
 
     if (adjusted_corr_obj$pass){
@@ -379,6 +397,7 @@ generate_pseudo_pop <- function(Y,
   result$pseudo_pop <- best_pseudo_pop
   result$adjusted_corr_results <- best_adjusted_corr_obj$corr_results
   result$original_corr_results <- original_corr_obj$corr_results
+  result$ks_stats <- best_ks_stats
   result$fcall <- fcall
   result$passed_covar_test <- adjusted_corr_obj$pass
   result$counter <- counter
@@ -388,6 +407,7 @@ generate_pseudo_pop <- function(Y,
   result$covariate_cols_name <- unlist(covariate_cols)
   result$ess <- ess
   result$ess_recommended <- ess_recommended
+
 
   end_time_gpp <- proc.time()